+++ 2023-csharp/year2023/Day03/Day03.definition.cs
@@ -2,7 +2,7 @@ namespace ofzza.aoc.year2023.day03;
 
 using ofzza.aoc.utils;
 
-public partial class Day03: ISolution<string[], int> {
+public partial class Day03: ISolution<string[], long> {
   public int SolutionYear { get; } = 2023;
 
   public int SolutionDay { get; } = 3;
@@ -19,24 +19,24 @@ public partial class Day03: ISolution<string[], int> {
     "...$.*....",
     ".664.598..",
   };
-  public List<SolutionExecution<string[], int>> Executions { get; } = new List<SolutionExecution<string[], int>>() {
+  public List<SolutionExecution<string[], long>> Executions { get; } = new List<SolutionExecution<string[], long>>() {
     // Part #1, Test
-    new SolutionExecution<string[], int>(1, Tag.Test) {
+    new SolutionExecution<string[], long>(1, Tag.Test) {
       InputValue = Day03.test,
       Expect = 4361
     },
     // Part #1, Solution
-    new SolutionExecution<string[], int>(1, Tag.Solution) {
+    new SolutionExecution<string[], long>(1, Tag.Solution) {
       InputValue = File.ReadAllText("./inputs/Day03/input.txt").Trim().Split("\n"),
       Expect = 527369
     },
     // Part #2, Test
-    new SolutionExecution<string[], int>(2, Tag.Test) {
+    new SolutionExecution<string[], long>(2, Tag.Test) {
       InputValue = Day03.test,
       Expect = 467835
     },
     // Part #2, Solution
-    new SolutionExecution<string[], int>(2, Tag.Solution) {
+    new SolutionExecution<string[], long>(2, Tag.Solution) {
       InputValue = File.ReadAllText("./inputs/Day03/input.txt").Trim().Split("\n"),
       Expect = 73074886
     }
